+++ .gitattributes
@@ -1,2 +1,2 @@
 *.mov filter=lfs diff=lfs merge=lfs -text
-osu!Radio/Background[[:space:]]Video/*.mov filter=lfs diff=lfs merge=lfs -text
+osuRadio/Background[[:space:]]Video/*.mov filter=lfs diff=lfs merge=lfs -text
