+++ .gitignore
@@ -182,3 +182,4 @@ updater.spec
 test/
 ffmpeg_bin/
 osuRadio/custom_songs
+test2/
