+++ osuRadio/main.py
@@ -1,613 +1,44 @@
-__version__ = "1.8.0"
+__version__ = "1.7.0"
 
 import sys
 import os
-import re
 import json
 import random
 import sqlite3
-import requests
 import zipfile
-import tarfile
 import tempfile
 import shutil
 import subprocess
 import tempfile
-import ffmpeg
-import platform
-import wave
-import hashlib
-from dateutil import parser as date_parser
 from packaging import version
 from pathlib import Path
-from mutagen.mp3 import MP3
 from time import monotonic
 from PySide6.QtCore import (
-    Qt, QUrl, QTimer, QThread, QMetaObject,
-    QPropertyAnimation, QEasingCurve, Property,
-    QSequentialAnimationGroup, QPauseAnimation, Signal,
+    Qt, QUrl, QTimer, QThread, QMetaObject, Signal,
     QEvent, QSize
 )
 from PySide6.QtGui import (
-    QIcon, QPixmap, QPainter, QColor,
-    QKeySequence, QShortcut, QCursor,
-    QGuiApplication
+    QIcon, QColor, QKeySequence, QShortcut, QCursor, QGuiApplication
 )
 from PySide6.QtWidgets import (
     QApplication, QMainWindow, QWidget, QLabel, QFileDialog,
     QHBoxLayout, QVBoxLayout, QListWidget, QListWidgetItem,
     QPushButton, QLineEdit, QSlider, QStyle,
     QDialog, QDialogButtonBox, QCheckBox, QComboBox,
-    QGraphicsOpacityEffect, QGraphicsColorizeEffect,
-    QMenu, QGridLayout, QToolTip, QSizePolicy,
+    QGraphicsOpacityEffect, QMenu, QGridLayout, QToolTip, QSizePolicy,
     QMessageBox, QProgressDialog, QScrollArea
 )
 from PySide6.QtMultimedia import QMediaPlayer, QAudioOutput, QVideoSink, QAudioFormat, QMediaDevices
 
-# Paths
-IS_WINDOWS = os.name == "nt"
-BASE_PATH = Path(getattr(sys, "frozen", False) and sys.executable or __file__).resolve().parent
-DATABASE_FILE = BASE_PATH / "songs.db"
-SETTINGS_FILE = BASE_PATH / "settings.json"
-CUSTOM_SONGS_PATH = BASE_PATH / "custom_songs"
-CUSTOM_SONGS_PATH.mkdir(exist_ok=True)
-EXPORT_STATE_FILE = BASE_PATH / "export_selected.json"
-
-if IS_WINDOWS:
-    import ctypes
-    try:
-        ctypes.windll.shcore.SetProcessDpiAwareness(1)
-    except Exception:
-        pass
-    user32 = ctypes.windll.user32
-else:
-    user32 = None
-
-WM_HOTKEY, MOD_NOREPEAT = 0x0312, 0x4000
-VK_MEDIA_PLAY_PAUSE, VK_MEDIA_NEXT_TRACK, VK_MEDIA_PREV_TRACK = 0xB3, 0xB0, 0xB1
-
-if getattr(sys, "frozen", False):
-    ASSETS_PATH = Path(sys._MEIPASS)
-    BASE_PATH = Path(sys.executable).parent  # For writing settings, DB
-else:
-    BASE_PATH = Path(__file__).parent
-    ASSETS_PATH = BASE_PATH
-
-if sys.platform == "darwin":
-    ICON_FILE = "Osu!RadioIcon.icns"
-elif sys.platform.startswith("linux"):
-    ICON_FILE = "Osu!RadioIcon.png"
-elif sys.platform.startswith("win"):
-    ICON_FILE = "Osu!RadioIcon.ico"
-else:
-    ICON_FILE = "Osu!RadioIcon.png"  # fallback
-    
-ICON_PATH = ASSETS_PATH / ICON_FILE
-IMG_PATH = ASSETS_PATH / "img"
-
-# FFmpeg bin setup
-def get_ffmpeg_bin_path():
-    base = Path(__file__).resolve().parent / "ffmpeg_bin"
-    system = platform.system().lower()
-    if system == "windows":
-        return base / "windows" / "bin" / "ffmpeg.exe"
-    elif system == "darwin":
-        return base / "macos" / "ffmpeg"
-    elif system == "linux":
-        return base / "linux" / "bin" / "ffmpeg"
-    else:
-        raise RuntimeError(f"Unsupported platform: {system}")
-
-ffmpeg_path = str(get_ffmpeg_bin_path())
-_original_popen = subprocess.Popen
-
-# Patch ffmpeg-python’s internal Popen to suppress terminal (for stream.run())
-def silent_popen(cmd, *args, **kwargs):
-    if sys.platform.startswith("win"):
-        si = subprocess.STARTUPINFO()
-        si.dwFlags |= subprocess.STARTF_USESHOWWINDOW
-        kwargs["startupinfo"] = si
-        kwargs["creationflags"] = subprocess.CREATE_NO_WINDOW
-    return _original_popen(cmd, *args, **kwargs)
-
-ffmpeg._run.Popen = silent_popen
-
-# Helper for silent subprocess.run settings
-def get_silent_subprocess_kwargs():
-    if sys.platform.startswith("win"):
-        si = subprocess.STARTUPINFO()
-        si.dwFlags |= subprocess.STARTF_USESHOWWINDOW
-        return {
-            "startupinfo": si,
-            "creationflags": subprocess.CREATE_NO_WINDOW,
-        }
-    return {}
-    
-def silent_global_popen(*args, **kwargs):
-    if sys.platform.startswith("win"):
-        si = subprocess.STARTUPINFO()
-        si.dwFlags |= subprocess.STARTF_USESHOWWINDOW
-        kwargs.setdefault("startupinfo", si)
-        kwargs.setdefault("creationflags", subprocess.CREATE_NO_WINDOW)
-    return _original_popen(*args, **kwargs)
-
-subprocess.Popen = silent_global_popen
-
-# Patch ffmpeg.run(...) to use subprocess.run with suppressed terminal
-def custom_run(*args, **kwargs):
-    cmd = [ffmpeg_path, *args[1:]]
-    subprocess_kwargs = {
-        "stdout": kwargs.get("stdout", subprocess.PIPE),
-        "stderr": kwargs.get("stderr", subprocess.PIPE),
-        "text": kwargs.get("text", False),
-        "check": kwargs.get("check", True),
-        **get_silent_subprocess_kwargs(),
-    }
-    return subprocess.run(cmd, **subprocess_kwargs)
-
-ffmpeg.run = custom_run
-
-# Utility functions
-def show_modal(msgbox: QMessageBox):
-    msgbox.setWindowModality(Qt.ApplicationModal)
-    msgbox.raise_()
-    msgbox.activateWindow()
-    msgbox.exec()
-
-def get_audio_path(song: dict) -> Path:
-    return Path(song["folder"]) / song["audio"]
-    
-def init_db():
-    with sqlite3.connect(DATABASE_FILE) as conn:
-        cursor = conn.cursor()
-        cursor.execute("""
-            CREATE TABLE IF NOT EXISTS songs (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                title TEXT,
-                artist TEXT,
-                mapper TEXT,
-                audio TEXT,
-                background TEXT,
-                length INTEGER,
-                osu_file TEXT,
-                folder TEXT,
-                UNIQUE(title, artist, mapper)
-            )""")
-        cursor.execute("""
-            CREATE TABLE IF NOT EXISTS metadata (
-                key TEXT PRIMARY KEY,
-                value TEXT
-            )""")
-        conn.commit()
-        
-def load_cache(folder: str) -> list | None:
-    if not DATABASE_FILE.exists():
-        return None
-    try:
-        with sqlite3.connect(DATABASE_FILE) as conn:
-            cursor = conn.cursor()
-            cursor.execute("SELECT value FROM metadata WHERE key = 'folder_mtime'")
-            mtime = cursor.fetchone()
-            if mtime and mtime[0] == str(os.path.getmtime(folder)):
-                cursor.execute("SELECT title, artist, mapper, audio, background, length, osu_file, folder FROM songs")
-                return [
-                    dict(zip(["title", "artist", "mapper", "audio", "background", "length", "osu_file", "folder"], row))
-                    for row in cursor.fetchall()
-                ]
-    except Exception as e:
-        print(f"[load_cache] Error: {e}")
-    return None
-    
-def save_cache(folder: str, maps: list):
-    init_db()
-    try:
-        with sqlite3.connect(DATABASE_FILE) as conn:
-            cursor = conn.cursor()
-            cursor.execute("BEGIN TRANSACTION")
-            for s in maps:
-                cursor.execute("""
-                    INSERT OR REPLACE INTO songs
-                    (title, artist, mapper, audio, background, length, osu_file, folder)
-                    VALUES (?, ?, ?, ?, ?, ?, ?, ?)""", (
-                        s.get("title"), s.get("artist"), s.get("mapper"),
-                        s.get("audio"), s.get("background"), s.get("length", 0),
-                        s.get("osu_file"), s.get("folder")
-                    ))
-            cursor.execute("INSERT OR REPLACE INTO metadata (key, value) VALUES (?, ?)",
-                           ('folder_mtime', str(os.path.getmtime(folder))))
-            conn.commit()
-    except Exception as e:
-        print(f"[save_cache] Error: {e}")
-        
-def read_osu_lines(path: str) -> list:
-    for enc in ("utf-8", "utf-8-sig", "cp1251", "latin-1"):
-        try:
-            with open(path, encoding=enc) as f:
-                return f.read().splitlines()
-        except UnicodeDecodeError:
-            continue
-    with open(path, encoding="utf-8", errors="ignore") as f:
-        return f.read().splitlines()
-        
-def get_audio_duration(file_path):
-    try:
-        probe = ffmpeg.probe(str(file_path))
-        duration = float(probe['format']['duration'])
-        return duration
-    except Exception as e:
-        print(f"Error getting duration: {e}")
-        return None
-
-def _hash_path(path: Path):
-    return hashlib.md5(str(path.resolve()).encode("utf-8")).hexdigest()[:10]
-
-def process_audio(input_file, speed=1.0, adjust_pitch=False, cache_dir=None):
-    input_path = Path(input_file)
-    if cache_dir is None:
-        cache_dir = Path(tempfile.gettempdir()) / "OsuRadioCache"
-    cache_dir.mkdir(exist_ok=True)
-
-    unique_id = _hash_path(input_path)
-    base_name = input_path.stem
-    suffix = f"{speed:.2f}x_pitch" if adjust_pitch else f"{speed:.2f}x"
-    output_file = cache_dir / f"{base_name}_{unique_id}_{suffix}.wav"
-
-    if output_file.exists():
-        return output_file, get_audio_duration(output_file)
-
-    print(f"[FFmpeg] Processing '{input_file}' → '{output_file}' with speed={speed}, pitch_adjust={adjust_pitch}")
-
-    stream = ffmpeg.input(str(input_file))
-
-    if speed == 1.0:
-        stream.output(str(output_file), format='wav', acodec="pcm_s16le", ac=2, ar=44100).run(overwrite_output=True, quiet=True)
-    else:
-        if adjust_pitch:
-            remaining = speed
-            while remaining < 0.5 or remaining > 2.0:
-                factor = 0.5 if remaining < 0.5 else 2.0
-                stream = stream.filter('atempo', factor)
-                remaining /= factor
-            stream = stream.filter('atempo', remaining)
-        else:
-            stream = stream.filter('rubberband', tempo=speed)
-
-        stream.output(str(output_file), acodec="pcm_s16le", ac=2, ar=44100).run(overwrite_output=True, quiet=True)
-
-    return output_file, get_audio_duration(output_file)
-        
-class PitchAdjustedPlayer:
-    def __init__(self, audio_output: QAudioOutput, parent=None):
-        self.player = QMediaPlayer(parent)
-        self.player.setAudioOutput(audio_output)
-        self._last_path = None
-        self.audio_output = audio_output
-        self.current_temp = None
-        self.last_temp = None
-        self.playback_rate = 1.0
-        self.preserve_pitch = True
-
-        self.last_start_ms = 0
-        self.was_playing_before_seek = False
-        self._pending_play = False
-
-        self.player.mediaStatusChanged.connect(self._start_after_load)
-
-    def _get_wav_duration_ms(self, path):
-        try:
-            if path.endswith(".wav"):
-                with wave.open(path, 'rb') as wf:
-                    frames = wf.getnframes()
-                    rate = wf.getframerate()
-                    return int((frames / rate) * 1000)
-            else:
-                duration = get_audio_duration(path)
-                return int(duration * 1000) if duration else 0
-        except Exception as e:
-            print(f"[Duration Error] Could not get duration for {path}: {e}")
-            return 0
-
-    def play(self, input_path: str, speed: float = 1.0, preserve_pitch: bool = True, start_ms: int = 0, force_play=False):
-        if (
-            self.player.mediaStatus() == QMediaPlayer.LoadedMedia
-            and self._last_path == input_path
-            and self.playback_rate == speed
-            and self.preserve_pitch == preserve_pitch
-        ):
-            print(f"[Resume] Resuming existing playback at {self.player.position()} ms")
-            if start_ms > 0:
-                self.player.setPosition(start_ms)
-            self.player.play()
-            return
-
-        self._last_path = input_path
-        self.preserve_pitch = preserve_pitch
-        self.playback_rate = speed
-        self.last_start_ms = start_ms
-        self.was_playing_before_seek = self.player.playbackState() == QMediaPlayer.PlayingState or force_play
-
-        if self.last_temp and self.last_temp.exists() and self.last_temp != self.current_temp:
-            try:
-                os.remove(self.last_temp)
-                print(f"[Cleanup] Deleted old cache: {self.last_temp}")
-            except Exception as e:
-                print(f"[Cleanup Error] Could not delete {self.last_temp}: {e}")
-
-        if preserve_pitch:
-            processed_file, _ = process_audio(input_path, speed=speed, adjust_pitch=True)
-            self.current_temp = Path(processed_file)
-            file_url = QUrl.fromLocalFile(str(processed_file))
-            self.player.setSource(file_url)
-            self._pending_play = True  # This triggers _start_after_load
-        else:
-            file_url = QUrl.fromLocalFile(str(input_path))
-            self.player.setSource(file_url)
-            self.player.setPlaybackRate(speed)
-            self.player.setPosition(start_ms)
-            self.player.play()
-            print(f"[QMediaPlayer] 🎵 Now playing: {file_url.toString()}")
-            self.current_temp = None
-            self._pending_play = False
-
-        self.last_temp = self.current_temp
-        self.last_duration = self._get_wav_duration_ms(file_url.toLocalFile())
-    
-    def _delayed_start(self):
-        if self.was_playing_before_seek:
-            self.player.play()
-        else:
-            self.player.pause()
-        self._pending_play = False
-
-    def _start_after_load(self, status):
-        if status == QMediaPlayer.LoadedMedia and self._pending_play:
-            print("[PitchPlayer] Media loaded, setting position")
-            self.player.setPosition(self.last_start_ms)
-            QTimer.singleShot(150, self._check_audio_after_load)
-
-    def _check_audio_after_load(self):
-        if self.was_playing_before_seek:
-            self.player.play()
-        else:
-            self.player.pause()
-
-        # Delay again to verify audio routing
-        QTimer.singleShot(500, self._verify_audio_available)
-
-    def _verify_audio_available(self):
-        if not self.player.isAvailable():
-            print("[PitchPlayer] No audio detected after load — retrying playback")
-            self.player.setPosition(self.last_start_ms)
-            self.player.play()
-
-    def stop(self):
-        self.player.stop()
-        if self.current_temp and os.path.exists(self.current_temp):
-            if self.current_temp != self.last_temp:
-                try:
-                    os.remove(self.current_temp)
-                    print(f"[Stop] Deleted current temp: {self.current_temp}")
-                except Exception as e:
-                    print(f"[Stop] Failed to delete {self.current_temp}: {e}")
-        self.current_temp = None
-
-class OsuParser:
-    @staticmethod
-    def parse(path: str) -> dict:
-        data = {
-            "audio": "", "title": "", "artist": "", "mapper": "",
-            "background": "", "length": 0,
-            "osu_file": path, "folder": str(Path(path).parent)
-        }
-        for line in read_osu_lines(path):
-            line = line.strip()
-            if m := re.match(r'audiofilename\s*:\s*(.+)', line, re.IGNORECASE):
-                data["audio"] = m.group(1).strip()
-            elif line.lower().startswith("title:"):
-                data["title"] = line.split(":", 1)[1].strip()
-            elif line.lower().startswith("artist:"):
-                data["artist"] = line.split(":", 1)[1].strip()
-            elif line.lower().startswith("creator:"):
-                data["mapper"] = line.split(":", 1)[1].strip()
-            elif line.startswith("0,0") and not data["background"]:
-                if bg := re.search(r'0,0,"([^"]+)"', line):
-                    data["background"] = bg.group(1)
-        try:
-            mp3_path = Path(data["folder"]) / data["audio"]
-            mp3 = MP3(str(mp3_path))
-            data["length"] = int(mp3.info.length * 1000)
-        except Exception:
-            pass
-        return data
-
-def check_for_update(current_version, skipped_versions=None, manual_check=False, include_prerelease=False):
-    url = "https://api.github.com/repos/Paraliyzedevo/osu-Radio/releases"
-    try:
-        releases = requests.get(url, timeout=5).json()
-        valid = []
-        current_parsed = version.parse(current_version.lstrip("v"))
-
-        for r in releases:
-            tag = r.get("tag_name", "").lstrip("v")
-            created = r.get("created_at")
-            is_prerelease = r.get("prerelease", False)
-
-            if not tag:
-                continue
-            if not manual_check and skipped_versions and tag in skipped_versions:
-                continue
-            if is_prerelease and not include_prerelease:
-                continue
-            if '+' in tag and not include_prerelease:
-                continue
-
-            try:
-                parsed_tag = version.parse(tag)
-                created_at = date_parser.parse(created)
-                valid.append((parsed_tag, is_prerelease, '+' in tag, created_at, r))
-            except Exception:
-                continue
-
-        if not valid:
-            return None, None
-
-        # Sort by: newest date, stable > pre, no local > local, then version (tie-breaker)
-        valid.sort(key=lambda v: (v[3], not v[1], not v[2], v[0]), reverse=True)
-        latest_ver, _, _, _, latest_data = valid[0]
-
-        if current_parsed < latest_ver:
-            return str(latest_ver), latest_data.get("assets", [])
-    except Exception as e:
-        print(f"[check_for_update] Failed: {e}")
-    return None, None
-
-def download_and_install_update(assets, latest_version, skipped_versions, settings_path, main_window=None):
-    platform = sys.platform
-    url = None
-
-    for asset in assets:
-        name = asset["name"].lower()
-        if platform.startswith("win") and name.endswith(".zip"):
-            url = asset["browser_download_url"]
-        elif platform == "darwin" and (name.endswith(".dmg") or name.endswith(".pkg")):
-            url = asset["browser_download_url"]
-        elif platform.startswith("linux") and name.endswith(".tar.gz"):
-            url = asset["browser_download_url"]
-        if url:
-            break
-
-    if not url:
-        QMessageBox.information(None, "Update", "No suitable update found.")
-        return
-
-    msg = QMessageBox()
-    msg.setWindowTitle("Update Available")
-    msg.setText(f"Version {latest_version} is available. Do you want to update?")
-    msg.setWindowIcon(QIcon(str(ICON_PATH)))
-    update_btn = msg.addButton("Update Now", QMessageBox.AcceptRole)
-    remind_btn = msg.addButton("Remind Me Later", QMessageBox.RejectRole)
-    skip_btn = msg.addButton("Skip This Version", QMessageBox.DestructiveRole)
-    show_modal(msg)
-
-    if msg.clickedButton() == skip_btn:
-        if latest_version not in skipped_versions:
-            skipped_versions.append(latest_version)
-        # Save immediately
-        try:
-            with open(settings_path, "r+") as f:
-                settings = json.load(f)
-                existing = settings.get("skipped_versions", [])
-                if latest_version not in existing:
-                    existing.append(latest_version)
-                    settings["skipped_versions"] = existing
-                    f.seek(0)
-                    json.dump(settings, f, indent=2)
-                    f.truncate()
-        except Exception as e:
-            print("Failed to save skipped version:", e)
-        return
-
-    if msg.clickedButton() != update_btn:
-        if main_window:
-            main_window.skip_downgrade_for_now = True
-        return
-
-    # Proceed with update
-    temp_dir = tempfile.mkdtemp()
-    file_path = os.path.join(temp_dir, os.path.basename(url))
-    with requests.get(url, stream=True) as r:
-        with open(file_path, "wb") as f:
-            total = int(r.headers.get('content-length', 0))
-            progress = QProgressDialog("Downloading update...", "Cancel", 0, total)
-            progress.setWindowModality(Qt.ApplicationModal)
-            progress.setWindowTitle("osu!Radio Updater")
-            progress.setWindowIcon(QIcon(str(ICON_PATH)))
-            progress.setMinimumWidth(400)
-            progress.show()
-
-            downloaded = 0
-            for chunk in r.iter_content(chunk_size=8192):
-                if not chunk:
-                    continue
-                f.write(chunk)
-                downloaded += len(chunk)
-                progress.setValue(downloaded)
-                QApplication.processEvents()
-                if progress.wasCanceled():
-                    QMessageBox.information(None, "Update Cancelled", "The update was cancelled.")
-                    return
-            progress.close()
-
-    extract_dir = tempfile.mkdtemp()
-    if file_path.endswith(".zip"):
-        with zipfile.ZipFile(file_path, "r") as zip_ref:
-            zip_ref.extractall(extract_dir)
-    elif file_path.endswith(".tar.gz"):
-        with tarfile.open(file_path, "r:gz") as tar_ref:
-            tar_ref.extractall(extract_dir)
-    else:
-        QMessageBox.information(
-            None,
-            "Manual Install Required",
-            f"The downloaded file ({os.path.basename(file_path)}) is a macOS installer.\n\n"
-            "1. Double-click the .dmg or .pkg file to open it.\n"
-            "2. Follow the on-screen instructions to complete the update.\n\n"
-            f"3. Close program (if you haven't done so) and replace all files with the new files downloaded on {file_path}"
-            f"File saved to: {file_path}"
-        )
-        return
-
-    # Look inside nested structure (like dist/osu!Radio/)
-    subdir = None
-    for root, dirs, files in os.walk(extract_dir):
-        for file in files:
-            if file.lower() == "osu!radio.exe":
-                subdir = root
-                break
-        if subdir:
-            break
-
-    if not subdir:
-        QMessageBox.warning(None, "Update Failed", "Could not find osu!Radio.exe in the extracted files.")
-        return
-
-    msg = QMessageBox()
-    msg.setIcon(QMessageBox.Information)
-    msg.setWindowTitle("Restarting")
-    msg.setText("osu!Radio will now restart with the update applied.")
-    msg.setStandardButtons(QMessageBox.Ok)
-    show_modal(msg)
-    ret = msg.result()
-
-    if ret == QMessageBox.Ok:
-        if getattr(sys, 'frozen', False):
-            if sys.platform.startswith("win"):
-                updater = os.path.join(sys._MEIPASS, "updater.exe")
-            elif sys.platform.startswith("linux"):
-               updater = os.path.join(sys._MEIPASS, "updater") 
-        else:
-            if sys.platform.startswith("win"):
-                updater = os.path.join(BASE_PATH, "updater.exe")
-            elif sys.platform.startswith("linux"):
-               updater = os.path.join(BASE_PATH, "updater")
-               
-        if sys.platform.startswith("linux"):
-            os.chmod(updater, 0o755)
-            
-        if sys.platform.startswith("win"):
-            exe = "osu!Radio.exe"
-        elif sys.platform.startswith("linux"):
-            exe = "osu!Radio"
-
-        temp_updater = tempfile.NamedTemporaryFile(delete=False, suffix=".exe").name
-        shutil.copy2(updater, temp_updater)
-
-        subprocess.Popen([
-            temp_updater,
-            subdir, str(BASE_PATH), exe, str(os.getpid())
-        ])
-        shutil.rmtree(temp_dir, ignore_errors=True)
-        sys.exit(0)
+from osuRadio.audio import PitchAdjustedPlayer, get_audio_duration
+from osuRadio.db import load_cache, save_cache, get_audio_path
+from osuRadio.parser import OsuParser
+from osuRadio.settings import SettingsDialog
+from osuRadio.ui import MarqueeLabel, BackgroundWidget
+from osuRadio.update import check_for_update, download_and_install_update
+from osuRadio.media_keys import update_media_key_listener
+from osuRadio.msg import show_modal
+from osuRadio.config import *
 
 class LibraryScanner(QThread):
     done = Signal(list)
@@ -657,255 +88,6 @@ class LibraryScanner(QThread):
 
         self.done.emit(library)
         print("[LibraryScanner] 'done' signal emitted.")
-
-class MarqueeLabel(QLabel):
-    def __init__(self, *args):
-        super().__init__(*args)
-        self._offset = 0
-        self._anim = QSequentialAnimationGroup(self)
-
-    def setText(self, txt):
-        super().setText(txt)
-        fm = self.fontMetrics()
-        text_w = fm.horizontalAdvance(txt)
-        avail = self.width()
-        self._anim.stop()
-        self._offset = 0
-
-        if text_w > avail:
-            span = text_w - avail + 20
-            a1 = QPropertyAnimation(self, b"offset", self)
-            a1.setStartValue(0)
-            a1.setEndValue(span)
-            a1.setDuration(span * 20)
-            a1.setEasingCurve(QEasingCurve.Linear)
-
-            p1 = QPauseAnimation(1000, self)
-
-            a2 = QPropertyAnimation(self, b"offset", self)
-            a2.setStartValue(span)
-            a2.setEndValue(0)
-            a2.setDuration(span * 20)
-            a2.setEasingCurve(QEasingCurve.Linear)
-
-            p2 = QPauseAnimation(1000, self)
-
-            self._anim.clear()
-            self._anim.addAnimation(a1)
-            self._anim.addAnimation(p1)
-            self._anim.addAnimation(a2)
-            self._anim.addAnimation(p2)
-            self._anim.setLoopCount(-1)
-            self._anim.start()
-        else:
-            self.update()
-
-    def paintEvent(self, ev):
-        painter = QPainter(self)
-        painter.setClipRect(self.rect())
-        painter.drawText(-self._offset, self.height() - 5, self.text())
-
-    def offset(self): return self._offset
-    def setOffset(self, v): self._offset = v; self.update()
-    offset = Property(int, offset, setOffset)
-
-class BackgroundWidget(QWidget):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._raw_frame = None
-        self._pixmap = QPixmap()
-        self._last_size = self.size()
-        self.effect = QGraphicsColorizeEffect(self)
-        self.effect.setStrength(1.0)
-        self.setGraphicsEffect(self.effect)
-
-        self._resize_timer = QTimer(self)
-        self._resize_timer.setSingleShot(True)
-        self._resize_timer.timeout.connect(self._rescale_pixmap)
-
-    def setFrame(self, frame):
-        self._raw_frame = frame
-        self._last_size = self.size()
-        img = frame.toImage()
-        self._pixmap = QPixmap.fromImage(img).scaled(
-            self.size(), Qt.KeepAspectRatioByExpanding, Qt.SmoothTransformation
-        )
-        self.update()
-
-    def resizeEvent(self, event):
-        if self._raw_frame and self.size() != self._last_size:
-            self._resize_timer.start(69)
-        super().resizeEvent(event)
-
-    def _rescale_pixmap(self):
-        if self._raw_frame:
-            self._last_size = self.size()
-            img = self._raw_frame.toImage()
-            self._pixmap = QPixmap.fromImage(img).scaled(
-                self.size(), Qt.KeepAspectRatioByExpanding, Qt.SmoothTransformation
-            )
-            self.update()
-
-    def paintEvent(self, ev):
-        painter = QPainter(self)
-        painter.setRenderHint(QPainter.SmoothPixmapTransform)
-
-        painter.drawPixmap(0, 0, self._pixmap)
-
-        # If the QGraphicsColorizeEffect is not active, apply hue manually
-        if not self.effect.isEnabled():
-            hue_color = self.effect.color()
-            overlay = QColor.fromHsv(hue_color.hue(), 255, 255)
-            overlay.setAlphaF(self.effect.strength())
-            painter.setCompositionMode(QPainter.CompositionMode_Multiply)
-            painter.fillRect(self.rect(), overlay)
-
-        painter.end()
-
-class SettingsDialog(QDialog):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.setWindowTitle("Settings")
-        self.setFixedSize(400, 350)
-        self.main = parent
-
-        layout = QVBoxLayout(self)
-
-        # Songs folder selection
-        self.folder_edit = QLineEdit(parent.osu_folder)
-        browse_btn = QPushButton("Browse…")
-        browse_btn.clicked.connect(self.browse_folder)
-
-        folder_layout = QHBoxLayout()
-        folder_layout.addWidget(QLabel("Songs Folder:"))
-        folder_layout.addWidget(self.folder_edit)
-        folder_layout.addWidget(browse_btn)
-        layout.addLayout(folder_layout)
-
-        # Toggles
-        self.light_mode_checkbox = QCheckBox("Light Mode")
-        self.video_checkbox = QCheckBox("Enable Background Video")
-        self.autoplay_checkbox = QCheckBox("Autoplay on Startup")
-        self.media_key_checkbox = QCheckBox("Enable Media Key Support")
-        self.pitch_checkbox = QCheckBox("Preserve Original Pitch (DT/NC)")
-        self.prerelease_checkbox = QCheckBox("Include Pre-release Updates")
-
-        self.light_mode_checkbox.setChecked(parent.light_mode)
-        self.video_checkbox.setChecked(parent.video_enabled)
-        self.autoplay_checkbox.setChecked(parent.autoplay)
-        self.media_key_checkbox.setChecked(parent.media_keys_enabled)
-        self.pitch_checkbox.setChecked(parent.preserve_pitch)
-        self.prerelease_checkbox.setChecked(parent.allow_prerelease)
-
-        for checkbox in (
-            self.light_mode_checkbox, self.video_checkbox, self.autoplay_checkbox,
-            self.media_key_checkbox, self.pitch_checkbox, self.prerelease_checkbox
-        ):
-            layout.addWidget(checkbox)
-
-        # Opacity slider
-        self.opacity_slider = QSlider(Qt.Horizontal)
-        self.opacity_slider.setRange(10, 100)
-        self.opacity_slider.setValue(int(parent.ui_opacity * 100))
-        self.opacity_slider.valueChanged.connect(
-            lambda v: parent.ui_effect.setOpacity(v / 100)
-        )
-
-        opacity_layout = QHBoxLayout()
-        opacity_layout.addWidget(QLabel("UI Opacity:"))
-        opacity_layout.addWidget(self.opacity_slider)
-        layout.addLayout(opacity_layout)
-
-        # Hue slider
-        self.hue_slider = QSlider(Qt.Horizontal)
-        self.hue_slider.setRange(0, 360)
-        self.hue_slider.setValue(parent.hue)
-        self.hue_slider.valueChanged.connect(
-            lambda v: parent.bg_widget.effect.setColor(QColor.fromHsv(v, 255, 255))
-        )
-
-        hue_layout = QHBoxLayout()
-        hue_layout.addWidget(QLabel("Hue:"))
-        hue_layout.addWidget(self.hue_slider)
-        layout.addLayout(hue_layout)
-
-        # Resolution dropdown
-        self.res_combo = QComboBox()
-        self.resolutions = {
-            "1920×1080": (1920, 1080),
-            "1280×720": (1280, 720),
-            "854×480": (854, 480),
-            "640×360": (640, 360),
-            "480×270": (480, 270),
-        }
-        self.res_combo.addItems(list(self.resolutions.keys()) + ["Custom Resolution"])
-        current_res = f"{parent.width()}×{parent.height()}"
-        if current_res in self.resolutions:
-            self.res_combo.setCurrentText(current_res)
-        else:
-            self.res_combo.setCurrentText("Custom Resolution")
-
-        res_layout = QHBoxLayout()
-        res_layout.addWidget(QLabel("Resolution:"))
-        res_layout.addWidget(self.res_combo)
-        layout.addLayout(res_layout)
-
-        # Update button
-        update_btn = QPushButton("Check for Updates")
-        update_btn.clicked.connect(lambda: parent.check_updates(manual=True))
-        layout.addWidget(update_btn)
-
-        # Buttons
-        buttons = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
-        buttons.accepted.connect(self.apply)
-        buttons.rejected.connect(self.reject)
-        layout.addWidget(buttons)
-
-        # Preview state backup
-        self._original_opacity = parent.ui_opacity
-        self._original_hue = parent.hue
-
-    def browse_folder(self):
-        folder = QFileDialog.getExistingDirectory(self, "Select osu! Songs Folder")
-        if folder:
-            self.folder_edit.setText(folder)
-
-    def apply(self):
-        folder = self.folder_edit.text()
-        light = self.light_mode_checkbox.isChecked()
-        opacity = self.opacity_slider.value() / 100
-        hue = self.hue_slider.value()
-        res = self.res_combo.currentText()
-        if res == "Custom Resolution":
-            w, h = self.main.width(), self.main.height()
-        else:
-            w, h = self.resolutions.get(res, (854, 480))
-        allow_resizing = res == "Custom Resolution"
-        video_on = self.video_checkbox.isChecked()
-        autoplay = self.autoplay_checkbox.isChecked()
-        media_keys = self.media_key_checkbox.isChecked()
-        preserve_pitch = self.pitch_checkbox.isChecked()
-        allow_prerelease = self.prerelease_checkbox.isChecked()
-
-        was_prerelease = self.main.allow_prerelease
-        self.main.apply_settings(
-            folder, light, opacity, w, h, hue,
-            video_on, autoplay, media_keys, preserve_pitch,
-            allow_prerelease, allow_resizing
-        )
-
-        if was_prerelease != allow_prerelease:
-            self.main.skip_downgrade_for_now = False
-            self.main.check_updates(manual=True)
-
-        self.accept()
-
-    def reject(self):
-        self.main.ui_effect.setOpacity(self._original_opacity)
-        self.main.bg_widget.effect.setColor(QColor.fromHsv(self._original_hue, 255, 255))
-        self.main.ui_opacity = self._original_opacity
-        self.main.hue = self._original_hue
-        super().reject()
         
 class MainWindow(QMainWindow):
     def __init__(self):  
@@ -1805,7 +987,7 @@ class MainWindow(QMainWindow):
         self.autoplay = autoplay
         if self.media_keys_enabled != media_keys:
             self.media_keys_enabled = media_keys
-            self.update_media_key_listener()
+            update_media_key_listener(self)
         self.preserve_pitch = preserve_pitch
         self.allow_prerelease = allow_prerelease
         if allow_resizing:
@@ -1889,14 +1071,6 @@ class MainWindow(QMainWindow):
             self.bg_widget.setGraphicsEffect(self.bg_widget.effect)
             self.bg_widget.effect.setEnabled(False)
             self.bg_widget.update()
-                
-    def update_media_key_listener(self):
-        try:
-            if self.media_key_listener:
-                self.media_key_listener.stop()
-                self.media_key_listener = None
-        except Exception as e:
-            print("Failed to stop media key listener:", e)
 
         if self.media_keys_enabled:
             try:
@@ -2308,9 +1482,11 @@ class MainWindow(QMainWindow):
             except Exception as e:
                 print(f"[Exit Cleanup] Cache delete failed: {e}")
 
-def run_app():
-    import sys
+if __name__ == "__main__":
     app = QApplication(sys.argv)
     window = MainWindow()
     window.show()
+    
+    update_media_key_listener(window)
+
     sys.exit(app.exec())
