+++ osuRadio/msg.py
@@ -1,0 +1,13 @@
+from PySide6.QtCore import (
+    Qt
+)
+from PySide6.QtWidgets import (
+    QMessageBox
+)
+from .config import *
+
+def show_modal(msgbox: QMessageBox):
+    msgbox.setWindowModality(Qt.ApplicationModal)
+    msgbox.raise_()
+    msgbox.activateWindow()
+    msgbox.exec()
