--- run.py
@@ -1,4 +1,0 @@
-from osuRadio.main import run_app
-
-if __name__ == "__main__":
-    run_app()
